Data Structes redesign and read ahead policy
Complete redesign of the cache data structures. It was a nightmare to read and debug the code. Now we use a a proper class to represent blocks and a special dictionary to maintain usage order and the blocks in the same place. Iterations are about 200us slower, but the code is more readable and maintanable.

Implementation of a read ahead policy that alows to configure how many contiguos blocks to read in advance when one is fetched from the device. This improves read speed (at least in FAT, LFS at this moment fails when RA>1)